fix: add shared code between dependencies, for now leveldb and inmemory. Signed-off-by: Francisco Javier Ribó Labrador <elribonazo@gmail.com>
