+++ packages/shared/README.md
@@ -1,0 +1,2 @@
+# @pluto-encrypted/shared
+Contains some shared dependencies for pluto encrypted.
