+++ packages/shared/rollup/rollup.mjs
@@ -1,0 +1,5 @@
+import nodePolyfills from "rollup-plugin-polyfill-node";
+
+import CreateConfig from "../../../rollup/index.mjs";
+
+export default CreateConfig(undefined, [nodePolyfills()]);
